distribute option

+++ skopy/commands/command_measure.py
@@ -1,3 +1,5 @@
+import os.path
+
 import click
 import pandas
 import sqlalchemy
@@ -5,24 +7,33 @@ import sqlalchemy.orm
 
 import skopy.command
 import skopy.feature
+import skopy.task
 
 
 @click.command("measure")
 @click.argument("metadata", nargs=1, type=click.Path(exists=True))
 @click.option("--database", default="sqlite:///measurements.sqlite")
+@click.option("--distribute", is_flag=True)
 @click.option("--verbose", is_flag=True)
-def command(metadata, database, verbose):
-    engine = sqlalchemy.create_engine(database, echo=verbose)
+def command(metadata, database, distribute, verbose):
+    if distribute:
+        pass
+    else:
+        engine = sqlalchemy.create_engine(database, echo=verbose)
 
-    skopy.feature.Base.metadata.drop_all(engine)
+        skopy.feature.Base.metadata.drop_all(engine)
 
-    skopy.feature.Base.metadata.create_all(engine)
+        skopy.feature.Base.metadata.create_all(engine)
 
-    session = sqlalchemy.orm.sessionmaker()
+        session = sqlalchemy.orm.sessionmaker()
 
-    session.configure(bind=engine)
+        session.configure(bind=engine)
 
-    session = session()
+        session = session()
+
+    directory = os.path.dirname(metadata)
+
+    directory = os.path.abspath(directory)
 
     metadata = pandas.read_csv(metadata)
 
@@ -30,8 +41,15 @@ def command(metadata, database, verbose):
 
     with progress as metadata:
         for pathname, mask in metadata:
-            image = skopy.feature.extract(pathname, mask)
+            pathname = os.path.join(directory, pathname)
 
-            session.add(image)
+            mask = os.path.join(directory, mask)
 
-    session.commit()
+            if distribute:
+                skopy.task.measure.delay(pathname, mask)
+            else:
+                image = skopy.feature.extract(pathname, mask)
+
+                session.add(image)
+
+                session.commit()
